requirements.txt
brother_ql-inventree
Flask
flask_bootstrap4
qrcode
pdf2image
Pillow==10.*
reportlab
requests
